requirements.txt
beautifulsoup4==4.12.3
Pillow==10.0.1
requests==2.32.3
